pyproject.toml
[project]
name = "goit-pythonweb-hw-10"
version = "0.1.0"
description = "API for storing and managing contacts"
authors = [
    {name = "Maryna-Korbet",email = "m22v86k@gmail.com"}
]
packages = [{ include = "src" }]
readme = "README.md"
requires-python = ">=3.12,<4.0"
dependencies = [
    "alembic (>=1.15.2,<2.0.0)",
    "sqlalchemy (>=2.0.40,<3.0.0)",
    "black (>=25.1.0,<26.0.0)",
    "asyncpg (>=0.30.0,<0.31.0)",
    "pydantic-settings (>=2.8.1,<3.0.0)",
    "pydantic (>=2.11.3,<3.0.0)",
    "uvicorn (>=0.34.0,<0.35.0)",
    "fastapi[standart] (>=0.115.12,<0.116.0)",
    "aiomysql (>=0.2.0,<0.3.0)",
    "pyjwt (>=2.10.1,<3.0.0)",
    "passlib[bcrypt] (>=1.7.4,<2.0.0)",
    "redis (>=5.2.1,<6.0.0)",
    "bcrypt (>=4.3.0,<5.0.0)",
    "email-validator (>=2.2.0,<3.0.0)",
    "python-multipart (>=0.0.20,<0.0.21)",
    "apscheduler (>=3.11.0,<4.0.0)",
    "requests (>=2.32.3,<3.0.0)",
    "slowapi (>=0.1.9,<0.2.0)",
    "fastapi-mail (>=1.4.2,<2.0.0)",
    "libgravatar (>=1.0.4,<2.0.0)",
    "cloudinary (>=1.44.0,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
